chore: refactor migrate and seed function
