+++ internal/db/reset/reset.go
@@ -73,7 +73,7 @@ func Run(ctx context.Context, version string, config pgconn.Config, fsys afero.F
 }
 
 func resetDatabase(ctx context.Context, version string, fsys afero.Fs, options ...func(*pgx.ConnConfig)) error {
-	fmt.Fprintln(os.Stderr, "Resetting local database"+getMessage(version))
+	fmt.Fprintln(os.Stderr, "Resetting local database"+toLogMessage(version))
 	if err := recreateDatabase(ctx, options...); err != nil {
 		return err
 	}
@@ -93,10 +93,10 @@ func resetDatabase(ctx context.Context, version string, fsys afero.Fs, options .
 		return err
 	}
 	defer conn.Close(context.Background())
-	return InitialiseDatabase(ctx, version, conn, fsys)
+	return apply.MigrateAndSeed(ctx, version, conn, fsys)
 }
 
-func getMessage(version string) string {
+func toLogMessage(version string) string {
 	if len(version) > 0 {
 		return " to version: " + version
 	}
@@ -110,13 +110,6 @@ func initDatabase(ctx context.Context, options ...func(*pgx.ConnConfig)) error {
 	}
 	defer conn.Close(context.Background())
 	return apply.BatchExecDDL(ctx, conn, strings.NewReader(utils.InitialSchemaSql))
-}
-
-func InitialiseDatabase(ctx context.Context, version string, conn *pgx.Conn, fsys afero.Fs) error {
-	if err := apply.MigrateDatabase(ctx, version, conn, fsys); err != nil {
-		return err
-	}
-	return SeedDatabase(ctx, conn, fsys)
 }
 
 // Recreate postgres database by connecting to template1
@@ -137,18 +130,6 @@ func recreateDatabase(ctx context.Context, options ...func(*pgx.ConnConfig)) err
 		},
 	}
 	return sql.ExecBatch(ctx, conn)
-}
-
-func SeedDatabase(ctx context.Context, conn *pgx.Conn, fsys afero.Fs) error {
-	seed, err := repair.NewMigrationFromFile(utils.SeedDataPath, fsys)
-	if errors.Is(err, os.ErrNotExist) {
-		return nil
-	} else if err != nil {
-		return err
-	}
-	fmt.Fprintln(os.Stderr, "Seeding data "+utils.Bold(utils.SeedDataPath)+"...")
-	// Batch seed commands, safe to use statement cache
-	return seed.ExecBatchWithCache(ctx, conn)
 }
 
 func DisconnectClients(ctx context.Context, conn *pgx.Conn) error {
@@ -241,7 +222,7 @@ func WaitForServiceReady(ctx context.Context, started []string) error {
 }
 
 func resetRemote(ctx context.Context, version string, config pgconn.Config, fsys afero.Fs, options ...func(*pgx.ConnConfig)) error {
-	fmt.Fprintln(os.Stderr, "Resetting remote database"+getMessage(version))
+	fmt.Fprintln(os.Stderr, "Resetting remote database"+toLogMessage(version))
 	conn, err := utils.ConnectRemotePostgres(ctx, config, options...)
 	if err != nil {
 		return err
@@ -264,7 +245,7 @@ func resetRemote(ctx context.Context, version string, config pgconn.Config, fsys
 	if err := migration.ExecBatch(ctx, conn); err != nil {
 		return err
 	}
-	return InitialiseDatabase(ctx, version, conn, fsys)
+	return apply.MigrateAndSeed(ctx, version, conn, fsys)
 }
 
 func ListSchemas(ctx context.Context, conn *pgx.Conn, exclude ...string) ([]string, error) {
