+++ internal/migration/apply/apply_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/supabase/cli/internal/migration/repair"
+	"github.com/supabase/cli/internal/testing/fstest"
 	"github.com/supabase/cli/internal/testing/pgtest"
 	"github.com/supabase/cli/internal/utils"
 )
@@ -45,22 +46,79 @@ func TestMigrateDatabase(t *testing.T) {
 		require.NoError(t, err)
 		defer mock.Close(ctx)
 		// Run test
-		err = MigrateDatabase(ctx, "", mock, fsys)
+		err = MigrateAndSeed(ctx, "", mock, fsys)
 		// Check error
 		assert.NoError(t, err)
 	})
 
 	t.Run("ignores empty local directory", func(t *testing.T) {
-		assert.NoError(t, MigrateDatabase(context.Background(), "", nil, afero.NewMemMapFs()))
+		assert.NoError(t, MigrateAndSeed(context.Background(), "", nil, afero.NewMemMapFs()))
 	})
 
 	t.Run("throws error on write failure", func(t *testing.T) {
 		// Setup in-memory fs
 		fsys := afero.NewMemMapFs()
 		// Run test
-		err := MigrateDatabase(context.Background(), "", nil, afero.NewReadOnlyFs(fsys))
+		err := MigrateAndSeed(context.Background(), "", nil, afero.NewReadOnlyFs(fsys))
 		// Check error
 		assert.ErrorIs(t, err, os.ErrPermission)
+	})
+}
+
+func TestSeedDatabase(t *testing.T) {
+	t.Run("seeds from file", func(t *testing.T) {
+		// Setup in-memory fs
+		fsys := afero.NewMemMapFs()
+		// Setup seed file
+		sql := "INSERT INTO employees(name) VALUES ('Alice')"
+		require.NoError(t, afero.WriteFile(fsys, utils.SeedDataPath, []byte(sql), 0644))
+		// Setup mock postgres
+		conn := pgtest.NewConn()
+		defer conn.Close(t)
+		conn.Query(sql).
+			Reply("INSERT 0 1")
+		// Connect to mock
+		ctx := context.Background()
+		mock, err := utils.ConnectLocalPostgres(ctx, pgconn.Config{Port: 5432}, conn.Intercept)
+		require.NoError(t, err)
+		defer mock.Close(ctx)
+		// Run test
+		assert.NoError(t, SeedDatabase(ctx, mock, fsys))
+	})
+
+	t.Run("ignores missing seed", func(t *testing.T) {
+		assert.NoError(t, SeedDatabase(context.Background(), nil, afero.NewMemMapFs()))
+	})
+
+	t.Run("throws error on read failure", func(t *testing.T) {
+		// Setup in-memory fs
+		fsys := &fstest.OpenErrorFs{DenyPath: utils.SeedDataPath}
+		// Run test
+		err := SeedDatabase(context.Background(), nil, fsys)
+		// Check error
+		assert.ErrorIs(t, err, os.ErrPermission)
+	})
+
+	t.Run("throws error on insert failure", func(t *testing.T) {
+		// Setup in-memory fs
+		fsys := afero.NewMemMapFs()
+		// Setup seed file
+		sql := "INSERT INTO employees(name) VALUES ('Alice')"
+		require.NoError(t, afero.WriteFile(fsys, utils.SeedDataPath, []byte(sql), 0644))
+		// Setup mock postgres
+		conn := pgtest.NewConn()
+		defer conn.Close(t)
+		conn.Query(sql).
+			ReplyError(pgerrcode.NotNullViolation, `null value in column "age" of relation "employees"`)
+		// Connect to mock
+		ctx := context.Background()
+		mock, err := utils.ConnectLocalPostgres(ctx, pgconn.Config{Port: 5432}, conn.Intercept)
+		require.NoError(t, err)
+		defer mock.Close(ctx)
+		// Run test
+		err = SeedDatabase(ctx, mock, fsys)
+		// Check error
+		assert.ErrorContains(t, err, `ERROR: null value in column "age" of relation "employees" (SQLSTATE 23502)`)
 	})
 }
